+++ src/modules/ekf2/EKF/bias_estimator/position_bias_estimator.hpp
@@ -39,7 +39,6 @@
 #define EKF_POSITION_BIAS_ESTIMATOR_HPP
 
 #include "bias_estimator.hpp"
-#include "common.h"
 
 class PositionBiasEstimator
 {
